requirements.txt
# Core dependencies
langchain-anthropic>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional: LangSmith for tracing
langsmith>=0.1.0
